+++ users.tf
@@ -23,6 +23,17 @@ data "aws_iam_policy_document" "ci_user" {
       "*",
     ]
   }
+  statement {
+    actions = [
+      "s3:ListBucket",
+      "s3:PutObject"
+    ]
+
+    resources = [
+      aws_s3_bucket.assets.arn,
+      format("%s/*", aws_s3_bucket.assets.arn)
+    ]
+  }
 }
 
 data "aws_iam_policy_document" "ci_user_lambda" {